Made actual use of the newly created class hierarchy to open IconSets.
Signals and Slots are being used to establish the communication
betweenMainWindow and Coordinator. IconSetOperator properly loads
the IconSet instance and offers getters to its access model classes,
which deliver data to the view.

Thereby the basic MVC structure has been prepared. Next steps
include the implementation of a mechanism to track the active IconSet
inside the Coordinator class.